+++ src/Settings.hpp
@@ -14,7 +14,7 @@ extern std::string MapName;
 extern std::string Key;
 extern int MaxPlayers;
 extern int ModsLoaded;
-extern int MaxModSize;
+extern long MaxModSize;
 extern bool Private;
 extern int MaxCars;
 extern bool Debug;
